Fixes #10673: Use a global event handler instead of individual ones on the input elements
Instead of registering a separate event handler for each input element (which breaks when the input element is replaced as done by the KeywordWidget on update), use a single handler registered on the container and use event bubbling instead.

No bonus points go to Javascript for not having a proper way to say "element in array".