re #11116: Fix image group type to ignore master image.
Also removed the old mechanism which preferred images with a scale of x140
which was used before the 2009 relaunch.

The image group type is used in XSLT to detuct the name of scaled images from
the image group name (folder), the desired scale (XxY) and the type:

    folder-XxY.type

Since the master image is not relevant at all for XSLT the new function tries
to ignore it for the type whenever this is possible.