Re #11354: Extract view behaviour for checkin from the view interface
The Checkin view exposes only two of the three states for semantic_change,
namely True and False, not None -- since those are the only ones that are needed right now.

In zeit.content.article we need the third state, and we need to be able to do something else
beside checkin in the view (namely, displaying a template).
Thus, we separate the checkin behaviour from "this is a view whose action it is to check in".