Fixes #11933: Always write updated xml value to the context
The XMLTree field was originally thought to be used on objects where
``xml`` is a property, thus replacing the value in the original tree
is enough (since the next get will read it from the tree again).

However, blocks have their XML node written to them, no property involved,
so in that use case that is not enough to make a change stick.

This has not been an issue before since the only place where it could have ocurred,
rawxml blocks on centerpages, don't redisplay the form after submit (it's inside a lightbox
which closes after submit), so they don't notice that they'd get the stale, old value.

To support this use case where ``xml`` is not a property, we now always write the changed value,
so it comes back from the object correctly (and replace it in the tree, if applicable, so the
change is persisted).