Re #12521: Trigger change after "generate keywords" (
This is a little duplicate work (since the keywords are generated on
the server and then updated in the DOM, it's a bit superfluous to tell
the server that keywords have changed), but this way error messages
are cleared immediately. The performance hit shouldn't be too severe,
and the user experience is better this way.