Fixes #12879: Remove temporary thumbnail files directly after the request ends
For thumbnails we create a LocalImage on the fly, which contains a
ZODB Blob to store the actual image data (in a temporary file).
Since that Blob never is part of a ZODB connection, we need to make
sure ourselves that it can be cleaned away when the request ends.