Fix test, it was using pytz in an unsupported way
http://pythonhosted.org/pytz/ says "Unfortunately using the tzinfo
argument of the standard datetime constructors *does not work* with
pytz for many timezones. It is [only] safe for timezones without
daylight saving transitions though, such as UTC".

This seems to have worked accidentally, but broke in pytz-2014.3,
now one really has to heed this and use localize() instead
(see also http://blog.qax.io/pytz-and-strange-timezones/).