MAINT: Simplify browser test setup
Instead of painstakingly avoiding site+interaction in the test setup,
just so that there is none when the testbrowser executes its request,
it's *much* easier to simply end any environment before actually
performing the request. As a bonus, we restore the previous environment
after the request is finished, so tests really should not have to take
any care about this at all anymore.

I'll remove the now-obsolete "with site/interaction" from all tests in
a separate commit, so this change is easier to see.