TMS-134: Resolve content without traversal if possible
This can be a massive performance improvement (in the order of 5x),
when in a non-zope.security context, i.e. where no __parent__ chain is
needed, and should not carry a penalty when it is (traversal: resolve
every folder from top to bottom; dynamic parents: resolve every folder
from bottom to top).

The code is inspired by monkey patches that have been used
successfully in zeit.web for a few years now.